Added hardcoded max_chunksize parameter
When run on a full dataset, multiprocess was crashing when communicating the results.
This is because the results were over 2GB in size, which was causing a problem when the code tried to pickle it.
The default chunksize (number of jobs given to a process at a time) is roughly number of genes / 4.
For a full dataset, this amount gaves results over 2GB per chunk.
By maxing out the chunksize at 100, this should hopefully overcome this problem.
max_chunksize should probably be a command line parameter, with the error caught and communicated properly.